refactor: put conn back to original location
