+++ config.py
@@ -1,6 +1,7 @@
 import os
 from typing import Optional
 
+import asyncpg
 import nextcord
 from dotenv.main import load_dotenv
 
@@ -16,6 +17,26 @@ _guild: Optional[nextcord.Guild] = None  # To be loaded on ready
 def guild() -> nextcord.Guild:
     assert _guild is not None
     return _guild
+
+
+_conn: Optional[asyncpg.Connection] = None  # The global connection to the database
+
+
+async def get_connection() -> asyncpg.Connection:
+    """Get a connection to the database.
+
+    If a connection has not yet been established or the current connection is closed, a new connection will be established.
+
+    Returns:
+        asyncpg.Connection: The connection to the database.
+    """
+    global _conn
+    if _conn is None or _conn.is_closed():
+        _conn = await asyncpg.connect(os.getenv("DATABASE_URL", ""), ssl="require")  # type: ignore
+    assert isinstance(
+        _conn, asyncpg.Connection
+    ), "A connection to the database could not be established."
+    return _conn
 
 
 # Google client configuration
