+++ database/person/person.py
@@ -1,6 +1,7 @@
 from typing import Iterable, Sequence, Set, Tuple
 
-from database import config as db_config, sql, sql_fetcher
+import config
+from database import sql, sql_fetcher
 
 
 class Person:
@@ -67,7 +68,7 @@ class Person:
         Returns:
             Sequence[Tuple[Person, float]]: A sequence of results where each result is a tuple of the person that matched as well as a similarity score between 0 and 1.
         """
-        conn = await db_config.get_connection()
+        conn = await config.get_connection()
         query = sql_fetcher.fetch("database", "person", "queries", "search_people.sql")
         return [(cls(*record[:-1]), record[-1]) for record in await conn.fetch(query, name)]
